requirements.txt
pandas>=1.3.4
numpy>=1.23.5
matplotlib>=3.5.0
streamlit==1.12.0
Pillow>=8.4.0